feat: extract model to separate classes
