+++ src/HassClient/Extensions/ServiceCollectionExtensions.cs
@@ -1,5 +1,6 @@
 ﻿using JoySoftware.HomeAssistant.Client;
 using JoySoftware.HomeAssistant.Helpers;
+using JoySoftware.HomeAssistant.Messages;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
 using Microsoft.Extensions.Logging;
