+++ src/HassClient/Messages/CommandMessage.cs
@@ -1,0 +1,10 @@
+﻿using System.Text.Json.Serialization;
+
+namespace JoySoftware.HomeAssistant.Messages
+{
+    public record CommandMessage : HassMessageBase
+    {
+        [JsonPropertyName("id")]
+        public int Id { get; set; }
+    }
+}
