+++ src/HassClient/Messages/GetConfigCommand.cs
@@ -1,0 +1,7 @@
+﻿namespace JoySoftware.HomeAssistant.Messages
+{
+    public record GetConfigCommand : CommandMessage
+    {
+        public GetConfigCommand() => Type = "get_config";
+    }
+}
