+++ src/HassClient/Messages/GetDevicesCommand.cs
@@ -1,0 +1,8 @@
+﻿namespace JoySoftware.HomeAssistant.Messages
+{
+    // {"type":"config/device_registry/list","id":25}
+    public record GetDevicesCommand : CommandMessage
+    {
+        public GetDevicesCommand() => Type = "config/device_registry/list";
+    }
+}
