+++ src/HassClient/Messages/GetServicesCommand.cs
@@ -1,0 +1,7 @@
+﻿namespace JoySoftware.HomeAssistant.Messages
+{
+    public record GetServicesCommand : CommandMessage
+    {
+        public GetServicesCommand() => Type = "get_services";
+    }
+}
