+++ src/HassClient/Messages/GetStatesCommand.cs
@@ -1,0 +1,7 @@
+﻿namespace JoySoftware.HomeAssistant.Messages
+{
+    public record GetStatesCommand : CommandMessage
+    {
+        public GetStatesCommand() => Type = "get_states";
+    }
+}
