+++ src/HassClient/Messages/HassPingCommand.cs
@@ -1,0 +1,7 @@
+﻿namespace JoySoftware.HomeAssistant.Messages
+{
+    public record HassPingCommand : CommandMessage
+    {
+        public HassPingCommand() => Type = "ping";
+    }
+}
