+++ src/HassClient/Model/HassServiceDomain.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections.Generic;
+
+namespace JoySoftware.HomeAssistant.Model
+{
+    public record HassServiceDomain
+    {
+        public string? Domain { get; init; }
+        
+        public IEnumerable<HassService>? Services { get; init; }
+    }
+}
