+++ src/HassClient/Model/HassServiceField.cs
@@ -1,0 +1,11 @@
+﻿namespace JoySoftware.HomeAssistant.Model
+{
+    public record HassServiceField
+    {
+        public string? Field { get; init; }
+        
+        public string? Description { get; init; }
+        
+        public object? Example { get; init; }
+    }
+}
